Add frontend store for persisted session/user data
- Before accessing a private route, first check that the session cookie
is set (user was previously logged in), then try to retrieve the
session data from the backend.
If either of these conditions fail, redirect to the login page.
- If we managed to get the data, use React Context to make it available
to all children components. Data flow is uni-directional (i.e. there's
no way to update the store from a child component) at the moment, but
this can be implemented if necessary.