Add user roles to session info API (backend)
To allow the frontend to render pages according to the current user's
role(s).

Roles are retrieved using SQL and made available to the controllers
as properties of `req.user`:
- isadmin
- ispetowner
- isfulltimecaretaker
- isparttimecaretaker

Note that these properties are case sensitive and are transformed into
their camelCase equivalents in the API, i.e. the frontend will see
isAdmin, isPetOwner, etc.